Added Stats for Emails

+++ app/Jobs/Referred/ReferredStatusChangeJob.php
@@ -28,21 +28,23 @@ class ReferredStatusChangeJob implements ShouldQueue
      */
     public function handle(): void
     {
-        switch($this->referred->reward_status && $this->referred->referrer){
-            case 'form_completed':
-                // If refereed has completed form, cancel eligibility emails
-                $this->referred->emailJobs()->whereIn('email_type', ['eligibility_email_2', 'eligibility_email_3', 'eligibility_email_4'])->where('email_sent', false)->delete();
-                break;
-            case 'not_interested':
-                // If referred is not interested, schedule nurture cycle email 1
-                $this->referred->emailJobs()->updateOrCreate(['email_type' => 'nurture_cycle_email_1', 'email_sent' => false], ['scheduled_date_time' => now()->addDays(2)]);
-                break;
-            case 'meeting_booked':
-                // If referred has booked meeting, cancel nurture cycle emails
-                $this->referred->emailJobs()->whereIn('email_type', ['nurture_cycle_email_1', 'nurture_cycle_email_2', 'nurture_cycle_email_3'])->where('email_sent', false)->delete();
-                break;
-            default:
-                break;
-        };
+        if($this->referred->referrer){
+            switch ($this->referred->reward_status) {
+                case 'form_completed':
+                    // If refereed has completed form, cancel eligibility emails
+                    $this->referred->emailJobs()->whereIn('email_type', ['eligibility_email_2', 'eligibility_email_3', 'eligibility_email_4'])->where('email_sent', false)->delete();
+                    break;
+                case 'not_interested':
+                    // If referred is not interested, schedule nurture cycle email 1
+                    $this->referred->emailJobs()->updateOrCreate(['email_type' => 'nurture_cycle_email_1', 'email_sent' => false], ['scheduled_date_time' => now()->addDays(2)]);
+                    break;
+                case 'meeting_booked':
+                    // If referred has booked meeting, cancel nurture cycle emails
+                    $this->referred->emailJobs()->whereIn('email_type', ['nurture_cycle_email_1', 'nurture_cycle_email_2', 'nurture_cycle_email_3'])->where('email_sent', false)->delete();
+                    break;
+                default:
+                    break;
+            };
+        }
     }
 }
